+++ ndd/base_estimator.py
@@ -32,11 +32,6 @@
 # DAMAGE.
 
 """Base estimator class from sklearn."""
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from builtins import (  # pylint: disable=redefined-builtin, unused-import
-    bytes, dict, int, list, object, range, str, ascii, chr, hex, input, next,
-    oct, open, pow, round, super, filter, map, zip)
 try:
     from inspect import signature  # pylint: disable=wrong-import-order
 except ImportError: