+++ ndd/nsb.py
@@ -1,13 +1,7 @@
 # -*- coding: utf-8 -*-
-# Copyright (C) 2016,2017 Simone Marsili
-# All rights reserved.
+# Author: Simone Marsili <simomarsili@gmail.com>
 # License: BSD 3 clause
 """Functions module."""
-from __future__ import (absolute_import, division, print_function,
-                        unicode_literals)
-from builtins import (  # pylint: disable=redefined-builtin, unused-import
-    bytes, dict, int, list, object, range, str, ascii, chr, hex, input, next,
-    oct, open, pow, round, super, filter, map, zip)
 import logging
 import numpy
 import ndd