+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Change log
 
 ## [Unreleased]
+### Changed
+- fixed NumericError for valid entropy estimation
 
 ## [0.9] - 2019-01-15
 ### Added