+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 # Change log
 
 ## [Unreleased]
+
+## [1.0] - 2019-01-15
 ### Changed
 - Python3 only
+- fixed NumericError for valid entropy estimation
 
 ## [0.9] - 2019-01-15
 ### Added